Conflicts:
	src/DataAcquisition/src/uk/ac/ox/zoo/seeg/abraid/mp/dataacquisition/Main.java
	src/DataAcquisition/src/uk/ac/ox/zoo/seeg/abraid/mp/dataacquisition/model/ModelRunManager.java
	src/DataAcquisition/src/uk/ac/ox/zoo/seeg/abraid/mp/dataacquisition/weightings/WeightingsCalculator.java
	src/DataAcquisition/test/uk/ac/ox/zoo/seeg/abraid/mp/dataacquisition/model/ModelRunRequesterIntegrationTest.java
	src/DataAcquisition/test/uk/ac/ox/zoo/seeg/abraid/mp/dataacquisition/weightings/WeightingsCalculatorIntegrationTest.java
	src/ModelWrapper/src/uk/ac/ox/zoo/seeg/abraid/mp/modelwrapper/model/data/OccurrenceDataWriterImpl.java

+++ src/Common/src/uk/ac/ox/zoo/seeg/abraid/mp/common/dao/HealthMapDiseaseDao.java
@@ -21,7 +21,7 @@ public interface HealthMapDiseaseDao {
      * @param id The ID.
      * @return The HealthMap disease, or null if not found.
      */
-    HealthMapDisease getById(Long id);
+    HealthMapDisease getById(Integer id);
 
     /**
      * Saves the specified HealthMap disease.
